execution schedule works. Tests separated.
