+++ pallets/execution-schedule/Cargo.toml
@@ -22,24 +22,24 @@ version = '1.1.0'
 [dependencies.frame-support]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+tag = 'v2.0.0-alpha.1'
 
 [dependencies.frame-system]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+tag = 'v2.0.0-alpha.1'
 
 [dependencies.sp-runtime]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+tag = 'v2.0.0-alpha.1'
 
 [dependencies.sp-std]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+tag = 'v2.0.0-alpha.1'
 
 [dev-dependencies]
 rand = "0.7.2"
-sp-core = { default_features = false, git = 'https://github.com/paritytech/substrate.git', rev = '3e651110aa06aa835790df63410a29676243fc54' }
-sp-io = { default_features = false, git = 'https://github.com/paritytech/substrate.git', rev = '3e651110aa06aa835790df63410a29676243fc54' }
+sp-core = { default_features = false, git = 'https://github.com/paritytech/substrate.git', tag = 'v2.0.0-alpha.1' }
+sp-io = { default_features = false, git = 'https://github.com/paritytech/substrate.git', tag = 'v2.0.0-alpha.1' }
